# docs/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Piums Platform API
  description: |
    API completa de la plataforma Piums - Marketplace para contratar artistas callejeros.
    
    ## Autenticación
    La mayoría de los endpoints requieren autenticación mediante JWT Bearer token.
    
    ```
    Authorization: Bearer <token>
    ```
    
    ## Ambientes
    - **Development**: http://localhost:3000
    - **Staging**: https://staging.piums.com
    - **Production**: https://piums.com
    
    ## Rate Limiting
    - **Login**: 5 requests / 15 minutos
    - **Register**: 3 requests / hora
    - **General**: 100 requests / 15 minutos
    
  version: 1.0.0
  contact:
    name: Piums API Support
    email: soporte@piums.io
    url: https://piums.com/docs
  license:
    name: Proprietary
    url: https://piums.com/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://staging.piums.com
    description: Staging server
  - url: https://piums.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint /api/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    
    # ========================================================================
    # Auth Schemas
    # ========================================================================
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - role
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePass123!"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [CLIENT, ARTIST, ADMIN]
          example: "CLIENT"
        phone:
          type: string
          example: "+34600000000"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
        expiresIn:
          type: string
          example: "15m"
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    
    # ========================================================================
    # User Schemas
    # ========================================================================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [CLIENT, ARTIST, ADMIN]
        avatar:
          type: string
          format: uri
          nullable: true
          example: "https://res.cloudinary.com/piums/image/upload/v1234567890/avatars/user.jpg"
        phone:
          type: string
          nullable: true
          example: "+34600000000"
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Registrar nuevo usuario
      description: Crea una cuenta nueva en la plataforma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Datos de registro inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: El email ya está registrado
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      description: Autenticarse con email y contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
        '429':
          description: Demasiados intentos de login
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar token
      description: Obtener un nuevo access token usando el refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresIn:
                    type: string
        '401':
          description: Refresh token inválido
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Cerrar sesión
      description: Invalidar el refresh token actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sesión cerrada
        '401':
          description: No autenticado
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Solicitar recuperación de contraseña
      description: Envía un email con link para resetear contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email enviado (siempre 200 por seguridad)
        '429':
          description: Demasiadas solicitudes
  /api/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Resetear contraseña
      description: Establecer nueva contraseña usando el token recibido por email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Contraseña actualizada
        '400':
          description: Token inválido o expirado
  /api/auth/change-password:
    post:
      tags:
        - Auth
      summary: Cambiar contraseña
      description: Cambiar contraseña del usuario autenticado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Contraseña cambiada
        '401':
          description: Contraseña actual incorrecta
  /api/auth/verify-email:
    post:
      tags:
        - Auth
      summary: Verificar email
      description: Confirmar dirección de email usando token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verificado
        '400':
          description: Token inválido
  /api/auth/resend-verification:
    post:
      tags:
        - Auth
      summary: Reenviar email de verificación
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Email reenviado
        '429':
          description: Demasiadas solicitudes

  # ==========================================================================
  # OAUTH
  # ==========================================================================
...